+++ lib/GenericsAPI/Utils/SampleServiceUtil.py
@@ -27,27 +27,30 @@ class SampleServiceUtil:
         self.scratch = config['scratch']
         self.token = config['KB_AUTH_TOKEN']
         self.srv_wiz_url = config['srv-wiz-url']
+        self.sample_url = config.get('kbase-endpoint') + '/sampleservice'
         self.dfu = DataFileUtil(self.callback_url)
-        self.sample_ser = SampleService(self.callback_url)
+        self.sample_ser = SampleService(self.sample_url)
 
         logging.basicConfig(format='%(created)s %(levelname)s: %(message)s',
                             level=logging.INFO)
 
     def get_sample_service_url(self):
 
-        payload = {
-            "method": "ServiceWizard.get_service_status",
-            "id": '',
-            "params": [{"module_name": "SampleService", "version": "dev"}],  # TODO: change to beta/release
-            "version": "1.1"
-        }
+        # payload = {
+        #     "method": "ServiceWizard.get_service_status",
+        #     "id": '',
+        #     "params": [{"module_name": "SampleService", "version": "dev"}],  # TODO: change to beta/release
+        #     "version": "1.1"
+        # }
 
-        sw_resp = requests.post(url=self.srv_wiz_url, data=json.dumps(payload))
-        wiz_resp = sw_resp.json()
-        if wiz_resp.get('error'):
-            raise RuntimeError("ServiceWizard Error - " + str(wiz_resp['error']))
+        # sw_resp = requests.post(url=self.srv_wiz_url, data=json.dumps(payload))
+        # wiz_resp = sw_resp.json()
+        # if wiz_resp.get('error'):
+        #     raise RuntimeError("ServiceWizard Error - " + str(wiz_resp['error']))
 
-        return wiz_resp['result'][0]['url']
+        # return wiz_resp['result'][0]['url']
+
+        return self.sample_url
 
     def get_sample(self, sample_id, version=None):