+++ RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # GenericsAPI release notes
 =========================================
+1.0.26
+Remove requirment of sample set for build_chemical_abundance_template method
+moving endpoint for SampleService from dynamic to core service
 
 1.0.25
 Adding functinoal profile data type support to the PCA app